Added in Player Guid initialization
The Player Controller http Post asks for a guid if an invalid one is provided it creates a new guid. The response body always contains the player guid.

I also refactored it so that the name can be updated with the post. Multiple users need to be handled on the front end if we want them.